backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
python-dotenv>=1.0.0
twitchio==2.6.0
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.50.0
